+++ src/components/order-stat/order-stat.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import styles from './order-stat.module.css';
+
+export default function OrderStat() {
+
+  return (
+    <div></div>
+  )
+}
