+++ src/pages/burger-container/burger-container.js
@@ -3,7 +3,7 @@ import styles from './burger-container.module.css';
 import BurgerIngredients from '../../components/burger-ingredients/burger-ingredients';
 import BurgerConstructor from '../../components/burger-constructor/burger-constructor';
 import Modal from '../../components/modal/modal';
-import OrderDetails from '../../components/order-details/order-details';
+import OrderInfo from '../../components/order-info/order-info';
 import { useSelector } from 'react-redux';
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
@@ -25,7 +25,7 @@ export default function BurgerContainer() {
   return (
     <main className={styles.box}>      
       {isModalOrder.success &&
-        <Modal children={<OrderDetails />} closeModal={closeModal}/>
+        <Modal children={<OrderInfo />} closeModal={closeModal}/>
       }
       <h1 className="text text_type_main-large mt-10 mb-5">
         Соберите бургер
