+++ src/pages/orders/orders.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import styles from './orders.module.css';
+import OrderList from '../../components/order-list/order-list';
+
+
+export default function Orders() {
+
+  return (
+    <div>
+      <OrderList />
+    </div>
+  );
+}
